switched vertices in glsl demo from f32 to u32

+++ src/screen/mod.rs
@@ -62,16 +62,16 @@ impl Screen {
       gl::Clear(gl::COLOR_BUFFER_BIT);
     }
     ////////////////////////////////////////////////////////////
-    let vertices: Vec<f32> = vec![256.0, 384.0,     255.0, 0.0,   0.0,
-                                  768.0, 384.0,    0.0,   255.0, 0.0,
-                                  512.0, 256.0,    0.0,   0.0, 255.0];
+    let vertices: Vec<u32> = vec![256, 384,     255, 0,   0,
+                                  768, 384,    0,   255, 0,
+                                  512, 256,    0,   0, 255];
     let mut vbo: gl::types::GLuint = 0;
     unsafe {
       gl::GenBuffers(1, &mut vbo);
       gl::BindBuffer(gl::ARRAY_BUFFER, vbo);
       gl::BufferData(
         gl::ARRAY_BUFFER,
-        (vertices.len() * std::mem::size_of::<f32>()) as gl::types::GLsizeiptr,
+        (vertices.len() * std::mem::size_of::<u32>()) as gl::types::GLsizeiptr,
         vertices.as_ptr() as *const gl::types::GLvoid,
         gl::STATIC_DRAW);
       gl::BindBuffer(gl::ARRAY_BUFFER, 0);
@@ -83,13 +83,13 @@ impl Screen {
       gl::BindVertexArray(vao);
       gl::BindBuffer(gl::ARRAY_BUFFER, vbo);
       gl::EnableVertexAttribArray(0);
-      gl::VertexAttribPointer(0, 2, gl::FLOAT, gl::FALSE,
-        (5 * std::mem::size_of::<f32>()) as gl::types::GLint,
+      gl::VertexAttribPointer(0, 2, gl::INT, gl::FALSE,
+        (5 * std::mem::size_of::<u32>()) as gl::types::GLint,
         std::ptr::null());
       gl::EnableVertexAttribArray(1);
-      gl::VertexAttribPointer(1, 3, gl::FLOAT, gl::FALSE,
-        (5 * std::mem::size_of::<f32>()) as gl::types::GLint,
-        (2 * std::mem::size_of::<f32>()) as *const gl::types::GLvoid);
+      gl::VertexAttribPointer(1, 3, gl::INT, gl::FALSE,
+        (5 * std::mem::size_of::<u32>()) as gl::types::GLint,
+        (2 * std::mem::size_of::<u32>()) as *const gl::types::GLvoid);
       gl::BindBuffer(gl::ARRAY_BUFFER, 0);
       gl::BindVertexArray(0);
     }
